Move ActionProviderBuilder to a specific file
